refactor(cart): extract CartKeypadController
Moves all keypad state and behaviour — active-row bookkeeping,
IME-visibility guard, slide-in/out animation, drag-to-dismiss touch
handling, and the seed/serialise CalculatorInputState helpers — into a
dedicated controller. The activity keeps the six android:onClick
reflection targets (numberEvent / decimalEvent / …) since
LayoutInflater resolves those against the Activity, but each one now
delegates through keypad.forwardKeypadAction. dispatchTouchEvent
forwards to keypad.handleTouchEvent, and drag-cancel dispatch is
threaded via a superDispatchTouchEvent bridge since Kotlin lambdas
can't reference super directly.

Net effect: CartActivity drops from ~750 to ~330 lines.
